ind2b_recommender/requirements.txt
# Core
python-dotenv
pandas
numpy
scipy

# MongoDB client
pymongo

# Machine Learning
scikit-learn

# API serving
fastapi
uvicorn

# Optional utilities
tqdm
requests
beautifulsoup4
